+++ CHANGELOG.md
@@ -15,7 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  -  When new row item is added and table page is full, jump to next page
  -  Active highlighting on hover and details button click to map filter results table
  -  Ability to delete landowner when editing a site
- -  Ability to delete sensor, data and nwis files in file tab
+ -  Ability to add sensor file, datafile and NWIS file in files tab
+ -  Ability to delete sensor, data and NWIS files in files tab
 
  
 ### Changed
@@ -30,6 +31,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  -  Fields not populating in details modals
  -  Unedited rows removed from reference datum table after edit
  -  Update counts in site page tabs on edit/add/delete
+ -  Validation error incorrectly shown when date changed in sensor retrieve modal
  -  UTC Preview not updating when minutes changed in Sensor add/edit
  -  Update counts in file tab and site file expansion panel when files added/deleted
  -  Date/time formatting and time zone conversion for all entities